lung segment

+++ segmentation/dataset/LungDataset.py
@@ -77,8 +77,8 @@ class LungDataset(Dataset):
         # if image.shape == 3:
             # import pdb; pdb.set_trace()
         # print(image.shape)
-        import pdb; pdb.set_trace()
-        return self.transform_image(np.array(image)), self.transform_label(np.array(mask)).long()
+        # import pdb; pdb.set_trace()
+        return self.transform_image(np.array(image)), self.transform_label(np.array(mask))
         
     def load_sample(self, batch_size = 4):
         list_imgs = []
